backend : update property rules

+++ backend/src/Controlleur/AnnonceControlleur.php
@@ -261,9 +261,15 @@ class AnnonceControlleur extends DefaultControlleur
     )]
     public function delete(int $id): void
     {
-        $this->isGranted(self::USER_ROLE);
-        if ($this->model->delete($id)) {
-            $this->jsonResponse("L'annonce a bien été delete");
+        $user = $this->isGranted(self::USER_ROLE);
+        $annonce = $this->model->find($id);
+        $userId = $annonce->getUserId();
+        if ($user["id"] == $userId) {
+            if ($this->model->delete($id)) {
+                $this->jsonResponse("L'annonce a bien été delete");
+            } else {
+                $this->jsonResponse("La suppréssion de l'annonce a échouer", 400);
+            }
         } else {
             $this->jsonResponse("La suppréssion de l'annonce a échouer", 400);
         }
